fix: build errors + obsolete-API warnings on Revit 2025+ host
Five build errors + six warnings reported by the CI Revit 2025 host
after Wave J pushed. All accumulated drift from earlier waves; no
behavioural changes — every fix preserves intent, just speaks the
modern API.

Errors fixed
1. BatchAssignCircuitsCommand.cs — pState was never declared.
   Wave H4 / Wave G timeline lost the `var pState = panels.Select(p =>
   new PanelState(p, circuitsByPanel)).ToList();` line during a refactor.
   Re-added at the top of the greedy assignment block.
2. BatchAssignCircuitsCommand.cs — ElectricalSystem.SelectPanel takes
   FamilyInstance in Revit 2024+, not the panel name string. Added
   PanelId field to Assignment so the apply pass resolves the
   FamilyInstance once via doc.GetElement(a.PanelId), then calls
   SelectPanel(panelInst). Also drops the redundant find-by-name walk
   for the panel-side ELC_PNL_CIRCUIT_GROUP_TXT stamp — direct id
   lookup is faster + correct.
3. AutoConduitDrop.cs — `fixtureEl.Host` doesn't exist on Element.
   Cast to FamilyInstance first: `var fi = fixtureEl as FamilyInstance;
   var wall = fi?.Host as Wall;` — null-safe, falls through cleanly
   when the fixture isn't a FamilyInstance or isn't wall-hosted.

Warnings fixed
- ElementId(int) → ElementId(long) in 7 call sites across
  ConduitAutoRouteCommand, ConduitConsolidator, CableScheduleBuilder,
  SlabPenetrationDetector. The (int) ctor is deprecated in Revit 2024
  and may be removed in a future release; the (long) overload is the
  documented replacement.
- SwapToManufacturerCommand: dropped unused `radiusFt` constant —
  only radiusFtSq is referenced for distance-squared comparisons.
- BatchAssignCircuitsCommand: removed orphaned static SafeName(
  FamilyInstance) at file scope — only the PanelState.SafeName variant
  is still called now that the find-by-name walk is gone.

SitePhotosTab.cs CS4014 warning is unrelated (base-branch code) and
left alone per scope discipline.

https://claude.ai/code

+++ StingTools/Commands/Electrical/Routing/CableScheduleBuilderCommand.cs
@@ -166,7 +166,7 @@ namespace StingTools.Commands.Electrical.Routing
             foreach (long id in allConduitIds)
             {
                 Element el = null;
-                try { el = doc.GetElement(new ElementId((int)id)); } catch { }
+                try { el = doc.GetElement(new ElementId((long)id)); } catch { }
                 if (el == null) continue;
                 var loc = el.Location as LocationCurve;
                 double mm = loc?.Curve?.Length * 304.8 ?? 0;
@@ -216,7 +216,7 @@ namespace StingTools.Commands.Electrical.Routing
             foreach (long id in allBoxIds)
             {
                 Element el = null;
-                try { el = doc.GetElement(new ElementId((int)id)); } catch { }
+                try { el = doc.GetElement(new ElementId((long)id)); } catch { }
                 if (el == null) continue;
 
                 string famName = "JB";
@@ -276,7 +276,7 @@ namespace StingTools.Commands.Electrical.Routing
                 {
                     try
                     {
-                        var el = doc.GetElement(new ElementId((int)id));
+                        var el = doc.GetElement(new ElementId((long)id));
                         var loc = el?.Location as LocationCurve;
                         if (loc?.Curve != null) total += loc.Curve.Length * 0.3048; // ft → m
                     }
