+++ StingTools/Commands/Electrical/Routing/ConduitAutoRouteCommand.cs
@@ -179,7 +179,7 @@ namespace StingTools.Commands.Electrical.Routing
                 {
                     if (c.RouteTrayIds == null) continue;
                     foreach (long lid in c.RouteTrayIds)
-                        jbConduitIds.Add(new ElementId((int)lid));
+                        jbConduitIds.Add(new ElementId((long)lid));
                 }
                 if (jbConduitIds.Count > 0)
                 {
@@ -248,7 +248,7 @@ namespace StingTools.Commands.Electrical.Routing
                 {
                     if (c.RouteTrayIds == null) continue;
                     foreach (long lid in c.RouteTrayIds)
-                        routedIds.Add(new ElementId((int)lid));
+                        routedIds.Add(new ElementId((long)lid));
                 }
                 if (routedIds.Count > 0)
                 {
